reservation cancel

+++ ci_intro/application/models/listsModel.php
@@ -24,9 +24,29 @@ class listsModel extends CI_Model
     }
 
     function delete_by_user($Did){
+
+        $x="yes";
+        $Rid="";
+        $query1=$this->db->query("SELECT * FROM res WHERE Did= ? ",array($Did));
+        foreach ($query1->result() as $row1) {
+          $Rid=$row1->Rid;
+        }
+        $data=array(
+      'Availability'=>$x
       
+      );
+      $this->db->where('Rid', $Rid);  
+      $res=$this->db->update('reservation', $data);  
+      if($res){
         $query=$this->db->query("DELETE FROM res WHERE Did= ? ",array($Did));
          return $query;
+      }
+      else{
+          return false;
+      }
+
+
+        
 
     }
